Clean up code for iPad fixes
